MeCloneMe_min_deploy/backend/requirements.txt
fastapi
uvicorn[standard]
jinja2==3.1.4
python-multipart
pydantic==2.9.2
sqlmodel==0.0.22
PyYAML
